Mypy fix

+++ src/cactus_runner/app/action.py
@@ -184,7 +184,7 @@ async def action_create_der_control(
         )
 
     randomize_seconds: int | None = resolved_parameters.get("randomizeStart_seconds", None)
-    ramp_time_seconds: float | None = resolved_parameters.get("ramp_time_seconds", None)
+    ramp_time_seconds: Decimal | None = resolved_parameters.get("ramp_time_seconds", None)
     energize: bool | None = resolved_parameters.get("opModEnergize", None)
     connect: bool | None = resolved_parameters.get("opModConnect", None)
     import_limit_watts: Decimal | None = resolved_parameters.get("opModImpLimW", None)
